Simplify the logic and refactor use-scheduled-save to use-throttle
