+++ packages/editor/src/components/autosave-monitor/test/index.js
@@ -10,7 +10,7 @@ import { AutosaveMonitor } from '../';
 
 const mockScheduleSave = jest.fn();
 const mockCancelSave = jest.fn();
-jest.mock( '../use-scheduled-save', () => {
+jest.mock( '../use-throttle', () => {
 	return () => ( {
 		scheduleSave: () => mockScheduleSave(),
 		cancelSave: () => mockCancelSave(),
@@ -23,70 +23,34 @@ describe( 'AutosaveMonitor', () => {
 		mockCancelSave.mockClear();
 	} );
 
-	it( 'should not schedule autosave on initial render', () => {
-		render( <AutosaveMonitor /> );
+	it( 'should schedule an autosave when dirty and saveable on initial render', () => {
+		render( <AutosaveMonitor isDirty={ true } isAutosaveable={ true } /> );
 		expect( mockCancelSave ).toHaveBeenCalledTimes( 0 );
-		expect( mockScheduleSave ).toHaveBeenCalledTimes( 0 );
+		expect( mockScheduleSave ).toHaveBeenCalledTimes( 1 );
 	} );
 
 	it( 'should schedule autosave when having become dirty and saveable', async () => {
-		const { rerender } = render( <AutosaveMonitor /> );
-		rerender(
-			<AutosaveMonitor isDirty={ true } isAutosaveable={ true } />
-		);
-
+		render( <AutosaveMonitor isDirty={ true } isAutosaveable={ true } /> );
 		expect( mockScheduleSave ).toHaveBeenCalledTimes( 1 );
 		expect( mockCancelSave ).toHaveBeenCalledTimes( 0 );
 	} );
 
-	it( 'should schedule autosave timer when edits reference changes', () => {
-		const beforeReference = [];
-		const afterReference = [];
-
-		const { rerender } = render(
-			<AutosaveMonitor
-				isDirty={ true }
-				isAutosaveable={ true }
-				editsReference={ beforeReference }
-			/>
-		);
-
-		expect( mockScheduleSave ).toHaveBeenCalledTimes( 0 );
-
-		rerender(
-			<AutosaveMonitor
-				isDirty={ true }
-				isAutosaveable={ true }
-				editsReference={ afterReference }
-			/>
-		);
-
-		expect( mockScheduleSave ).toHaveBeenCalledTimes( 1 );
-	} );
-
 	it( 'should stop autosave timer when the autosave is up to date', () => {
-		const { rerender } = render( <AutosaveMonitor /> );
-		rerender(
-			<AutosaveMonitor isDirty={ false } isAutosaveable={ true } />
-		);
+		render( <AutosaveMonitor isDirty={ false } isAutosaveable={ true } /> );
 
 		expect( mockScheduleSave ).toHaveBeenCalledTimes( 0 );
 		expect( mockCancelSave ).toHaveBeenCalledTimes( 1 );
 	} );
 
 	it( 'should stop autosave timer when having become dirty but not autosaveable', () => {
-		const { rerender } = render( <AutosaveMonitor /> );
-		rerender(
-			<AutosaveMonitor isDirty={ true } isAutosaveable={ false } />
-		);
+		render( <AutosaveMonitor isDirty={ true } isAutosaveable={ false } /> );
 
 		expect( mockScheduleSave ).toHaveBeenCalledTimes( 0 );
 		expect( mockCancelSave ).toHaveBeenCalledTimes( 1 );
 	} );
 
 	it( 'should stop autosave timer when having become not dirty', () => {
-		const { rerender } = render( <AutosaveMonitor /> );
-		rerender(
+		const { rerender } = render(
 			<AutosaveMonitor isDirty={ true } isAutosaveable={ true } />
 		);
 		rerender(
@@ -97,8 +61,7 @@ describe( 'AutosaveMonitor', () => {
 	} );
 
 	it( 'should stop autosave timer when having become not autosaveable', () => {
-		const { rerender } = render( <AutosaveMonitor /> );
-		rerender(
+		const { rerender } = render(
 			<AutosaveMonitor isDirty={ true } isAutosaveable={ true } />
 		);
 		rerender(
@@ -106,51 +69,6 @@ describe( 'AutosaveMonitor', () => {
 		);
 
 		expect( mockCancelSave ).toHaveBeenCalledTimes( 1 );
-	} );
-
-	it( 'should avoid scheduling autosave if still dirty but already autosaved for edits', () => {
-		const { rerender } = render( <AutosaveMonitor /> );
-
-		// Explanation: When a published post is autosaved, it's still in a
-		// dirty state since the edits are not saved to the post until the
-		// user clicks "Update". To avoid recurring autosaves, ensure that
-		// an edit has occurred since the last autosave had completed.
-
-		const beforeReference = [];
-		const afterReference = [];
-
-		// A post is non-dirty while autosave is in-flight.
-		rerender(
-			<AutosaveMonitor
-				isDirty={ false }
-				isAutosaving={ true }
-				isAutosaveable={ true }
-				editsReference={ beforeReference }
-			/>
-		);
-		expect( mockCancelSave ).toHaveBeenCalledTimes( 1 );
-		rerender(
-			<AutosaveMonitor
-				isDirty={ true }
-				isAutosaving={ false }
-				isAutosaveable={ true }
-				editsReference={ beforeReference }
-			/>
-		);
-
-		expect( mockCancelSave ).toHaveBeenCalledTimes( 2 );
-
-		// Once edit occurs after autosave, resume scheduling.
-		rerender(
-			<AutosaveMonitor
-				isDirty={ true }
-				isAutosaving={ false }
-				isAutosaveable={ true }
-				editsReference={ afterReference }
-			/>
-		);
-
-		expect( mockScheduleSave ).toHaveBeenCalledTimes( 1 );
 	} );
 
 	it( 'should render nothing', () => {
